Optimisation and publishing
Edit the CSS and JS bundles to have extensions and tell IIS to let us
handle those paths. This lets us output the bundles to those requests and
because it has an extension IIS will be smart enough to compress it for
us.

Update publishing profile to generate an installable zip instead of just
dumping the files where the old site was hosted.