Begin making this project a little bit organized.
There is now an interface for the forum scraper just in case we want to
continue scraping both. The Task that used to create and launch the
scraper now instantiates a manager class instead and calls its Run()
method which does the work.

At the moment we share a database context but I'm not sure I like it.
Passing the accounts in and using a new context to save doesn't work
because you need to call SaveChanges on the context the parent object
(the Account) is on and an object can only belong to a single change
tracker.