No Migrations until we figure out what we need.
Added extension methods to build view models. If building threads
manually you need to specify that it should instantiate the posts view
models itself. When instantiating the view models for a collection of
posts the models for the relevant threads will be created automatically.

Some of those .ToList() can probably be removed.
Some places may still be broken, not sure if scraping works but assume
it will.